Added possibility to use doubles in xml file + fixed GainOptimization double histograms

+++ DQMUtils/DQMHistogramBase.h
@@ -50,7 +50,7 @@ class DQMHistogramBase
      * \brief Book histograms
      * \param theDetectorStructure : Container of the Detector structure
      */
-    virtual void book(TFile *outputFile, const DetectorContainer &theDetectorStructure, std::map<std::string, uint32_t> pSettingsMap) = 0;
+    virtual void book(TFile *outputFile, const DetectorContainer &theDetectorStructure, std::map<std::string, double> pSettingsMap) = 0;
 
     /*!
      * \brief Book histograms
@@ -137,7 +137,7 @@ class DQMHistogramBase
 	    }
     }
 
-  auto findValue (const std::map<std::string, uint32_t>& pSettingsMap, const char* name)
+  auto findValue (const std::map<std::string, double>& pSettingsMap, const char* name)
   {
     auto setting = pSettingsMap.find(name);
     return ((setting != std::end(pSettingsMap)) ? setting->second : 0);
